requirements.txt
Appium-Python-Client==3.2.1
selenium==4.23.0
pytest==8.0.0
classutilities==0.2.1
pytest_assertions==0.5.0
python-dotenv==1.0.1
sqlmodel==0.0.21
httpx==0.27.0
pydantic-settings==2.4.0
PyMySQL==1.1.1
aiomysql==0.2.0
SQLAlchemy==2.0.32